Create SudoKube
John, a research scholar / Professor / Puzzle solver wants your help in publishing his work on SudoKube on his online blog for his followers and students.

A SudoKube is a mixture of Rubics cube and Sudoku. A SudoKube has exactly 6 appearances of every digit from 1 to 9 across the cube, whereas Rubics cube has 6 different colours.

As John wants to publish his work in text /document form (no video) he's concerned about how he would depict the step-by-step work of rotation in 2D form. Following are the notions and concepts John follows:

1) The six faces of the cube are named FRONT, BACK, UP, DOWN, LEFT, and RIGHT respectively.

2) Just like a Rubics cube which moves in 90 and 180 degrees in both clockwise and anti-clockwise directions, so can the SudoKube.

3) Any given face of the cube is a 3x3 square matrix whose indices are denoted by (0,0) to (2,2). The diagram below illustrates the same.

4) An elementary move is denoted in the following fashion:

• If a given face is rotated 90 degrees clockwise about the axis passing from the center of the face to the center of the cube, the move is denoted by the first letter of the name of the face.

• If the rotation is anticlockwise by 90 degrees, the letter is followed by an apostrophe (').

• If the rotation is by 180 degrees, the letter is followed by a 2.

Above diagram displays the indices of the matrix on the faces

John wants to test his notations on you. He has given you the initial position of the SudoKube and he has given you a set of operations to be performed on the SudoKube basis of his notation. After applying all the operations, the final SudoKube state should be the same as what John expects. Your task is to apply the operations and print the final SudoKube state.